+++ projects/ngx-material/src/lib/ngx-dialog/directives/dialog-content-template.directive.ts
@@ -1,9 +1,13 @@
-import {Directive, TemplateRef} from '@angular/core';
+import {Directive, TemplateRef, ViewContainerRef} from '@angular/core';
+import {RenderSource} from "@consensus-labs/ngx-tools";
 
 @Directive({selector: '[dialog-content-tmpl]'})
-export class DialogContentTemplateDirective {
+export class DialogContentTemplateDirective implements RenderSource {
 
-  constructor(public template: TemplateRef<void>) {
+  constructor(
+    public readonly template: TemplateRef<void>,
+    public readonly viewContainer: ViewContainerRef
+  ) {
   }
 
 }