+++ projects/ngx-material/src/lib/ngx-dialog/directives/dialog-footer-template.directive.ts
@@ -1,7 +1,11 @@
-import {Directive, TemplateRef} from '@angular/core';
+import {Directive, TemplateRef, ViewContainerRef} from '@angular/core';
+import {RenderSource} from "@consensus-labs/ngx-tools";
 
 @Directive({selector: '[dialog-footer-tmpl]'})
-export class DialogFooterTemplateDirective {
-  constructor(public template: TemplateRef<void>) {
+export class DialogFooterTemplateDirective implements RenderSource {
+  constructor(
+    public readonly template: TemplateRef<void>,
+    public readonly viewContainer: ViewContainerRef
+  ) {
   }
 }