+++ projects/ngx-material/src/lib/ngx-dialog/models/template-dialog-context.ts
@@ -1,6 +1,6 @@
 import {Observable, of, ReplaySubject, Subscription} from "rxjs";
-import {AnyTemplate, OverlayToken, SimpleTemplateRendering, TemplateRendering} from "@consensus-labs/ngx-tools";
-import {ViewContainerRef} from "@angular/core";
+import {OverlayToken, Rendering, RenderSource, TemplateRendering} from "@consensus-labs/ngx-tools";
+import {Injector} from "@angular/core";
 
 export interface TemplateDialogOptions {
   header$: Observable<string>;
@@ -24,10 +24,6 @@ export abstract class TemplateDialogContext implements TemplateDialogOptions {
 
 export class TemplateDialogInstance extends TemplateDialogContext {
 
-  get injector() {
-    return this.viewContainer.injector
-  }
-
   _content$ = new ReplaySubject<TemplateRendering>();
   content$ = this._content$.asObservable();
   _footer$ = new ReplaySubject<TemplateRendering | undefined>();
@@ -38,38 +34,42 @@ export class TemplateDialogInstance extends TemplateDialogContext {
   private footer?: TemplateRendering;
 
   constructor(
-    private viewContainer: ViewContainerRef,
     private token: OverlayToken,
     options: TemplateDialogOptions,
-    contentTemplate: Observable<AnyTemplate> | AnyTemplate,
-    footerTemplate?: Observable<AnyTemplate | undefined> | AnyTemplate,
+    contentTemplate: Observable<RenderSource> | RenderSource,
+    footerTemplate?: Observable<RenderSource | undefined> | RenderSource,
+    public readonly injector?: Injector,
   ) {
     super(options, token.zIndex);
     this.token.handleEscape(() => this.close());
 
     const content$ = contentTemplate instanceof Observable ? contentTemplate : of(contentTemplate);
-    this.sub.add(content$.subscribe(template => {
+    this.sub.add(content$.subscribe(source => {
       this.content?.dispose();
-      this.content = new SimpleTemplateRendering(this.viewContainer, template);
+      this.content = Rendering.FromSource.Static(source);
       this._content$.next(this.content);
     }));
 
     if (!footerTemplate) return;
 
     const footer$ = footerTemplate instanceof Observable ? footerTemplate : of(footerTemplate);
-    this.sub.add(footer$.subscribe(template => {
-      this.content?.dispose();
-      if (!template) return;
-      this.content = new SimpleTemplateRendering(this.viewContainer, template);
-      this._content$.next(this.content);
+    this.sub.add(footer$.subscribe(source => {
+      this.footer?.dispose();
+      if (!source) return;
+      this.footer = Rendering.FromSource.Static(source);
+      this._footer$.next(this.footer);
     }));
   }
 
   dispose() {
     this.token.dispose();
     this.sub.unsubscribe();
-    this.content?.dispose();
-    this.footer?.dispose();
+
+    // Delay until after animation ends
+    setTimeout(() => {
+      this.content?.dispose();
+      this.footer?.dispose();
+    }, 200);
   }
 
   private closeCallback?: () => any;