+++ projects/ngx-material/src/lib/ngx-dialog/services/dialog-manager.service.ts
@@ -1,8 +1,8 @@
-import {Injectable, Injector, ViewContainerRef} from "@angular/core";
+import {Injectable, Injector} from "@angular/core";
 import {Scheduler} from "@consensus-labs/rxjs-tools";
 import {Observable} from "rxjs";
 import {DialogInstance} from "../models/dialog-context";
-import {AnyTemplate, OverlayService} from "@consensus-labs/ngx-tools";
+import {OverlayService, RenderSource} from "@consensus-labs/ngx-tools";
 import {TemplateDialogInstance, TemplateDialogOptions} from "../models/template-dialog-context";
 import {StaticDialogInstance, StaticDialogOptions} from "../models/static-dialog-context";
 
@@ -18,14 +18,13 @@ export class DialogManagerService {
   }
 
   createDialog(
-    viewContainer: ViewContainerRef,
-    contentTemplate: Observable<AnyTemplate>|AnyTemplate,
-    footerTemplate: Observable<AnyTemplate|undefined>|AnyTemplate|undefined,
+    injector: Injector|undefined,
+    contentTemplate: Observable<RenderSource>|RenderSource,
+    footerTemplate: Observable<RenderSource|undefined>|RenderSource|undefined,
     options: TemplateDialogOptions,
   ): TemplateDialogInstance {
     const token = this.overlayService.pushOverlay();
     const instance = new TemplateDialogInstance(
-      viewContainer,
       token,
       options,
       contentTemplate,