+++ projects/ngx-tools/src/lib/rendering/index.ts
@@ -1,5 +1,6 @@
 
 export * from './models/template-rendering'
-export * from './models/data-template'
+export * from './models/render-source'
 export * from './directives/render-outlet.directive'
+export * from './directives/render-source.directive'
 export * from './ngx-rendering.module'