+++ projects/ngx-tools/src/lib/rendering/models/render-source.ts
@@ -1,0 +1,6 @@
+import {TemplateRef, ViewContainerRef} from "@angular/core";
+
+export interface RenderSource<T = void> {
+  readonly template: TemplateRef<T>;
+  readonly viewContainer: ViewContainerRef;
+}