+++ projects/ngx-tools/src/lib/rendering/ngx-rendering.module.ts
@@ -1,9 +1,10 @@
 import {NgModule} from '@angular/core';
 import {RenderOutletDirective} from "./directives/render-outlet.directive";
+import {RenderSourceDirective} from "./directives/render-source.directive";
 
 @NgModule({
-  exports: [RenderOutletDirective],
-  declarations: [RenderOutletDirective],
+  declarations: [RenderOutletDirective, RenderSourceDirective],
+  exports: [RenderOutletDirective, RenderSourceDirective],
 })
 export class NgxRenderingModule {
 }